refactor(1013): extract OperationRegistry to src/utils/ (SC-001, position 13)
Move the ~460-line OperationRegistry class from MistHelper.py to
src/utils/operation_registry.py, keeping a re-export in MistHelper so
MistHelper.OperationRegistry keeps resolving for the 10 external test
references and the 4 internal callsites in the systematic-test harness.

FR-006 remediation: introduce module-level _natural_option_sort_key +
_sort_options helpers so safe_options / interactive_safe_options /
unsafe_options no longer duplicate the historical
`float(x.replace("a", ".1"))` sort key.

FR-008 remediation: drop the # type: ignore[no-untyped-def, type-arg,
no-any-return] markers by typing the option-list helpers as
Iterable[str] -> list[str] and the registry entries as _OptionEntry.
