+++ MistHelper.py
@@ -247,6 +247,9 @@ from src.ui.display_utils import (  # pylint: disable=unused-import
 from src.ui.interactive_display_utils import (  # pylint: disable=unused-import
     InteractiveDisplayUtils,  # noqa: F401  # Cat B (1013 SC-001 position 10) -- re-export for callers at 17392/17393/17394/17395
 )
+from src.utils.operation_registry import (  # pylint: disable=unused-import
+    OperationRegistry,  # noqa: F401  # Cat B (1013 SC-001 position 13) -- re-export for menu safety classification
+)
 from src.wan_hub_group_manager import WanHubGroupNumberManager  # Import WAN hub group number manager for hub routing
 from src.wan_vpn_builder import WanVpnBuilder  # Import WAN VPN configuration builder
 from src.websocket.commands import MacTableCommand  # Import WebSocket show-MAC-table command handler
@@ -17578,470 +17581,7 @@ menu_actions = {
 # NOTE: TelemetryEmitter has been extracted to src/analytics/telemetry_emitter.py (issue #1013 SC-001 position 9)
 
 
-# ============================================================================
-# OPERATION REGISTRY CLASS
-# ============================================================================
-class OperationRegistry:
-    """Centralised classification of all menu operations.
-
-    Replaces the inline ``unsafe_options`` and ``interactive_read_only_options``
-    dictionaries with a single authoritative source.  Each entry maps a menu
-    option string to a dict with ``category`` and optional ``skip_reason``.
-
-    Categories
-    ----------
-    safe             Automated GET -- runs in ``--test``
-    interactive_safe Read-only but needs site/device -- runs in ``--testinteractive``
-    destructive      Modifies state -- always skipped
-    wip              Work in progress -- unstable
-    resource_intensive  Takes >1 h or hits rate limits
-    websocket        Requires WebSocket + interactive selection
-    continuous_loop  Never terminates without user stop
-    interactive      Needs user input not automatable
-    """
-
-    _REGISTRY: dict = {  # type: ignore[type-arg]
-        # --- control --------------------------------------------------------
-        "0": {"category": "interactive", "skip_reason": "Exit option"},
-        # --- resource intensive ---------------------------------------------
-        "19": {
-            "category": "resource_intensive",
-            "skip_reason": "Port-level statistics - extremely resource intensive (8+ hours)",
-        },
-        "59": {
-            "category": "resource_intensive",
-            "skip_reason": "Site configurations - hits API rate limits after 7+ hours",
-        },
-        # --- websocket ------------------------------------------------------
-        "102": {
-            "category": "websocket",
-            "skip_reason": "WebSocket ping - requires interactive site and device selection",
-        },
-        "103": {
-            "category": "websocket",
-            "skip_reason": "WebSocket traceroute - requires interactive site and device selection",
-        },
-        "104": {
-            "category": "websocket",
-            "skip_reason": "WebSocket release DHCP - requires interactive site and device selection",
-        },
-        "105": {
-            "category": "websocket",
-            "skip_reason": "WebSocket cable test - requires interactive site and device selection",
-        },
-        "118": {
-            "category": "websocket",
-            "skip_reason": "WebSocket bounce port - requires interactive site and device selection",
-        },
-        "119": {
-            "category": "websocket",
-            "skip_reason": "WebSocket ARP - requires interactive site and device selection",
-        },
-        "120": {
-            "category": "websocket",
-            "skip_reason": "WebSocket service ping - requires interactive site and device selection",
-        },
-        "121": {
-            "category": "websocket",
-            "skip_reason": "WebSocket ARP command - requires interactive site and device selection",
-        },
-        # --- interactive (needs user input, not automatable) -----------------
-        "134": {
-            "category": "interactive",
-            "skip_reason": "Packet capture - requires interactive configuration and site selection",
-        },
-        "135": {
-            "category": "interactive",
-            "skip_reason": "Packet capture - requires interactive configuration and MxEdge ID",
-        },
-        "136": {"category": "interactive", "skip_reason": "MSP export - requires interactive MSP selection"},
-        "137": {
-            "category": "interactive",
-            "skip_reason": "Firmware upgrade status - requires interactive scope selection",
-        },
-        "138": {"category": "interactive", "skip_reason": "CSV comparison - requires interactive file selection"},
-        "139": {
-            "category": "interactive",
-            "skip_reason": "Marvis troubleshooting - requires interactive option selection",
-        },
-        "92": {"category": "interactive_safe", "skip_reason": "Interactive site selection"},
-        "93": {"category": "interactive_safe", "skip_reason": "Interactive site inventory browser"},
-        "94": {"category": "interactive_safe", "skip_reason": "Interactive device stats viewer"},
-        "95": {"category": "interactive_safe", "skip_reason": "Interactive device tests viewer"},
-        "96": {"category": "interactive_safe", "skip_reason": "Interactive device config viewer"},
-        "140": {"category": "interactive", "skip_reason": "Interactive CLI shell session"},
-        "141": {"category": "interactive", "skip_reason": "Interactive TUI API browser - keyboard navigation required"},
-        "148": {
-            "category": "interactive",
-            "skip_reason": "WLAN RADIUS timer management - requires interactive site selection",
-        },
-        "149": {
-            "category": "interactive",
-            "skip_reason": "Requires interactive site selection for WAN2 variable configuration",
-        },
-        "150": {
-            "category": "interactive",
-            "skip_reason": "Requires interactive template selection for configuration extraction",
-        },
-        "142": {
-            "category": "interactive",
-            "skip_reason": "Maps Manager - requires interactive Dash web server and browser",
-        },
-        "143": {"category": "interactive", "skip_reason": "Requires interactive login with email/password credentials"},
-        "144": {"category": "interactive", "skip_reason": "MSP Inventory Export - requires MSP privileges via --login"},
-        # --- interactive_safe (read-only, need site/device, automatable) -----
-        "62": {"category": "interactive_safe", "skip_reason": "Requires site selection"},
-        "65": {"category": "interactive_safe", "skip_reason": "Requires site selection"},
-        "60": {"category": "interactive_safe", "skip_reason": "Requires site selection"},
-        "61": {"category": "interactive_safe", "skip_reason": "Requires site selection"},
-        "63": {"category": "interactive_safe", "skip_reason": "Requires site and device selection"},
-        "64": {"category": "interactive_safe", "skip_reason": "Requires site selection"},
-        "69": {"category": "interactive_safe", "skip_reason": "Requires site selection"},
-        "66": {"category": "interactive_safe", "skip_reason": "Requires site selection"},
-        "67": {"category": "interactive_safe", "skip_reason": "Requires site selection"},
-        "68": {"category": "interactive_safe", "skip_reason": "Requires site selection"},
-        "73": {"category": "interactive_safe", "skip_reason": "Requires site selection"},
-        "74": {"category": "interactive_safe", "skip_reason": "Requires site selection"},
-        "75": {"category": "interactive_safe", "skip_reason": "Requires site selection"},
-        "76": {"category": "interactive_safe", "skip_reason": "Requires site selection"},
-        "77": {"category": "interactive_safe", "skip_reason": "Requires site selection"},
-        "78": {"category": "interactive_safe", "skip_reason": "Requires site and device selection"},
-        "79": {"category": "interactive_safe", "skip_reason": "Requires site and client selection"},
-        # --- continuous loop ------------------------------------------------
-        "151": {"category": "continuous_loop", "skip_reason": "Continuous loop operation"},
-        "152": {"category": "continuous_loop", "skip_reason": "Continuous data collection loop"},
-        # --- resource intensive (file/support) ------------------------------
-        "100": {
-            "category": "resource_intensive",
-            "skip_reason": "File processing operation - potentially resource intensive",
-        },
-        "101": {
-            "category": "resource_intensive",
-            "skip_reason": "Support package generation - potentially resource intensive",
-        },
-        # --- long-running exports (52-week time windows) --------------------
-        "97": {"category": "resource_intensive", "skip_reason": "52-week device events export - long-running"},
-        "98": {"category": "resource_intensive", "skip_reason": "52-week audit logs export - long-running"},
-        "99": {"category": "resource_intensive", "skip_reason": "All-site gateway configs export - long-running"},
-        # --- destructive ----------------------------------------------------
-        "154": {"category": "destructive", "skip_reason": "DESTRUCTIVE: AP firmware upgrade operation"},
-        "158": {"category": "destructive", "skip_reason": "DESTRUCTIVE: Device reboot operation"},
-        "161": {"category": "destructive", "skip_reason": "DESTRUCTIVE: Virtual chassis conversion"},
-        "162": {"category": "destructive", "skip_reason": "DESTRUCTIVE: Virtual chassis conversion - bulk operation"},
-        "175": {
-            "category": "destructive",
-            "skip_reason": "Enhanced SSH Command Runner - requires interactive host and command input",
-        },
-        "176": {
-            "category": "destructive",
-            "skip_reason": "SSH Runner by gateway template - requires interactive template and command input",
-        },
-        "155": {"category": "destructive", "skip_reason": "DESTRUCTIVE: Switch firmware upgrade operation"},
-        "156": {"category": "destructive", "skip_reason": "DESTRUCTIVE: SSR firmware upgrade operation"},
-        "163": {"category": "destructive", "skip_reason": "DESTRUCTIVE: Updates gateway templates with WAN2 variable"},
-        "164": {"category": "destructive", "skip_reason": "DESTRUCTIVE: Applies gateway template configuration"},
-        "171": {"category": "destructive", "skip_reason": "DESTRUCTIVE: Creates 137 test sites from CSV"},
-        "172": {"category": "destructive", "skip_reason": "DESTRUCTIVE: Creates country-specific RF templates"},
-        "173": {"category": "destructive", "skip_reason": "DESTRUCTIVE: Creates device profiles for AP models"},
-        "174": {"category": "destructive", "skip_reason": "DESTRUCTIVE: Assigns APs to device profiles"},
-        "165": {"category": "destructive", "skip_reason": "DESTRUCTIVE: Clones gateway templates by state/country"},
-        "166": {"category": "destructive", "skip_reason": "DESTRUCTIVE: Configures WAN probe override on templates"},
-        "167": {"category": "destructive", "skip_reason": "DESTRUCTIVE: Configures WAN probe on device port overrides"},
-        "157": {"category": "destructive", "skip_reason": "DESTRUCTIVE: Org-level AP firmware upgrade"},
-        "168": {"category": "destructive", "skip_reason": "DESTRUCTIVE: Site Auto-Upgrade Configuration"},
-        "169": {"category": "destructive", "skip_reason": "DESTRUCTIVE: Site Analytics Configuration"},
-        "170": {
-            "category": "destructive",
-            "skip_reason": "DESTRUCTIVE: Bulk RADIUS WLAN Configuration - modifies WLAN auth settings",
-        },
-        # Device Utility Commands (123-157)
-        "123": {
-            "category": "websocket",
-            "skip_reason": "WebSocket traceroute - interactive site/device",
-        },
-        "106": {
-            "category": "websocket",
-            "skip_reason": "WebSocket OSPF neighbors - interactive gateway",
-        },
-        "107": {
-            "category": "websocket",
-            "skip_reason": "WebSocket OSPF interfaces - interactive gateway",
-        },
-        "108": {
-            "category": "websocket",
-            "skip_reason": "WebSocket OSPF database - interactive gateway",
-        },
-        "109": {
-            "category": "websocket",
-            "skip_reason": "WebSocket OSPF summary - interactive gateway",
-        },
-        "110": {
-            "category": "websocket",
-            "skip_reason": "WebSocket show sessions - interactive gateway",
-        },
-        "111": {
-            "category": "websocket",
-            "skip_reason": "WebSocket show service path - interactive gateway",
-        },
-        "112": {
-            "category": "websocket",
-            "skip_reason": "WebSocket BGP summary - interactive device",
-        },
-        "113": {
-            "category": "websocket",
-            "skip_reason": "WebSocket ARP table - interactive device",
-        },
-        "114": {
-            "category": "websocket",
-            "skip_reason": "WebSocket DHCP leases - interactive device",
-        },
-        "115": {
-            "category": "websocket",
-            "skip_reason": "WebSocket 802.1X table - interactive switch",
-        },
-        "116": {
-            "category": "websocket",
-            "skip_reason": "WebSocket EVPN database - interactive device",
-        },
-        "117": {
-            "category": "websocket",
-            "skip_reason": "WebSocket DNS resolution - interactive gateway",
-        },
-        "124": {
-            "category": "interactive",
-            "skip_reason": "Monitor traffic streaming - interactive port + Ctrl+C",
-        },
-        "125": {
-            "category": "interactive",
-            "skip_reason": "Run top streaming - interactive device + Ctrl+C",
-        },
-        "128": {
-            "category": "interactive",
-            "skip_reason": "Locate device - interactive device selection",
-        },
-        "129": {
-            "category": "interactive",
-            "skip_reason": "Unlocate device - interactive device selection",
-        },
-        "159": {
-            "category": "destructive",
-            "skip_reason": "DESTRUCTIVE: Bounce port - disrupts traffic",
-        },
-        "122": {
-            "category": "websocket",
-            "skip_reason": "WebSocket cable test - interactive switch/port",
-        },
-        "160": {
-            "category": "destructive",
-            "skip_reason": "DESTRUCTIVE: Reprovision - pushes fresh config",
-        },
-        "130": {
-            "category": "interactive",
-            "skip_reason": "Re-adopt device - interactive switch selection",
-        },
-        "131": {
-            "category": "interactive",
-            "skip_reason": "ZTP password - interactive device selection",
-        },
-        "132": {
-            "category": "interactive",
-            "skip_reason": "Config CLI commands - interactive switch",
-        },
-        "133": {
-            "category": "interactive",
-            "skip_reason": "Upload support file - interactive device/type",
-        },
-        "177": {
-            "category": "destructive",
-            "skip_reason": "DESTRUCTIVE: Clear ARP cache",
-        },
-        "178": {
-            "category": "destructive",
-            "skip_reason": "DESTRUCTIVE: Clear BGP routes",
-        },
-        "179": {
-            "category": "destructive",
-            "skip_reason": "DESTRUCTIVE: Clear session",
-        },
-        "180": {
-            "category": "destructive",
-            "skip_reason": "DESTRUCTIVE: Clear MAC table",
-        },
-        "181": {
-            "category": "destructive",
-            "skip_reason": "DESTRUCTIVE: Clear BPDU errors",
-        },
-        "182": {
-            "category": "destructive",
-            "skip_reason": "DESTRUCTIVE: Clear learned MACs from port",
-        },
-        "183": {
-            "category": "destructive",
-            "skip_reason": "DESTRUCTIVE: Clear policy hit count",
-        },
-        "184": {
-            "category": "destructive",
-            "skip_reason": "DESTRUCTIVE: Release DHCP lease",
-        },
-        "185": {
-            "category": "destructive",
-            "skip_reason": "DESTRUCTIVE: Release SSR DHCP lease",
-        },
-        "126": {
-            "category": "interactive",
-            "skip_reason": "Poll switch stats - interactive switch",
-        },
-        "127": {
-            "category": "interactive",
-            "skip_reason": "Create device snapshot - interactive switch",
-        },
-        "26": {"category": "safe"},
-        "145": {
-            "category": "interactive",
-            "skip_reason": "Interactive multi-phase workflow with write-capable phases",
-        },
-        "89": {"category": "interactive_safe"},
-        "90": {"category": "interactive_safe"},
-        "91": {"category": "interactive_safe"},
-        "146": {
-            "category": "interactive",
-            "skip_reason": "Interactive VPN pod management with API writes",
-        },
-        "147": {
-            "category": "interactive",
-            "skip_reason": "Interactive VPN builder with API writes",
-        },
-        "153": {
-            "category": "resource_intensive",
-            "skip_reason": "Bulk org data collection - runs 57 API calls, resource intensive",
-        },
-        "70": {"category": "interactive_safe", "skip_reason": "Requires site selection"},
-        "71": {"category": "interactive_safe", "skip_reason": "Requires site selection"},
-        "72": {"category": "interactive_safe", "skip_reason": "Requires site selection"},
-        "88": {"category": "interactive_safe", "skip_reason": "Requires AP model selection"},
-        "25": {"category": "safe"},
-        "196": {"category": "safe"},
-        "186": {"category": "destructive", "skip_reason": "DANGEROUS: Deletes all generated cache CSV files"},
-        "58": {"category": "safe"},
-        "187": {"category": "destructive", "skip_reason": "DESTRUCTIVE: Creates config objects in destination org"},
-        # Site Stats, Metrics & Channel Planning (178-185)
-        "80": {"category": "interactive_safe", "skip_reason": "Requires site selection"},
-        "81": {"category": "interactive_safe", "skip_reason": "Requires site selection"},
-        "82": {"category": "interactive_safe", "skip_reason": "Requires site selection"},
-        "83": {"category": "interactive_safe", "skip_reason": "Requires site selection"},
-        "84": {"category": "interactive_safe", "skip_reason": "Requires site selection"},
-        "85": {"category": "interactive_safe", "skip_reason": "Requires site selection"},
-        "86": {"category": "interactive_safe", "skip_reason": "Requires site selection"},
-        "23": {"category": "safe"},
-        # HA Gateway Cluster Info (186)
-        "87": {"category": "interactive_safe", "skip_reason": "Requires site selection"},
-        # Org Device Inventory Summary (187) - fully automated, no user input required
-        "13": {"category": "safe"},
-    }
-
-    # Wave 1 deterministic baseline map used by routing guardrail tests.
-    # This remains intentionally small and representative for compliance checks.
-    WAVE1_ENTRY_ROUTING_BASELINE: dict[str, str] = {
-        "102": "websocket",
-        "62": "interactive_safe",
-        "97": "resource_intensive",
-        "120": "websocket",
-        "154": "destructive",
-        "158": "destructive",
-        "156": "destructive",
-        "141": "interactive",
-        "26": "safe",
-        "58": "safe",
-        "187": "destructive",
-    }
-
-    # Wave 1 deterministic safety-boundary baseline used by classification guardrails.
-    WAVE1_SAFETY_CLASSIFICATION_BASELINE: dict[str, list[str]] = {
-        "safe_true": ["26", "58", "9999"],
-        "safe_false": ["120", "154", "158", "155", "156", "187"],
-        "interactive_safe_true": ["62"],
-        "interactive_safe_false": ["154", "141", "187"],
-        "destructive_markers": ["154", "158", "155", "156", "171", "159", "187"],
-    }
-
-    # Categories that are safe for --test (fully automated, no user input)
-    SAFE_CATEGORIES = frozenset({"safe"})
-    # Categories that are safe for --testinteractive (need site but automatable)
-    INTERACTIVE_SAFE_CATEGORIES = frozenset({"interactive_safe"})
-    # Categories always skipped
-    SKIP_CATEGORIES = frozenset(
-        {
-            "destructive",
-            "wip",
-            "resource_intensive",
-            "websocket",
-            "continuous_loop",
-            "interactive",
-        }
-    )
-
-    @classmethod
-    def get(cls, option: str) -> dict:  # type: ignore[type-arg]
-        """Return classification for *option*, defaulting to safe with warning."""
-        entry = cls._REGISTRY.get(str(option))
-        if entry is not None:
-            return entry  # type: ignore[no-any-return]
-        logging.warning("OperationRegistry: option %s not registered, defaulting to safe", option)
-        return {"category": "safe"}
-
-    @classmethod
-    def is_safe(cls, option: str) -> bool:
-        """True if *option* can run in ``--test`` mode."""
-        return cls.get(option)["category"] in cls.SAFE_CATEGORIES
-
-    @classmethod
-    def is_interactive_safe(cls, option: str) -> bool:
-        """True if *option* can run in ``--testinteractive`` mode."""
-        return cls.get(option)["category"] in cls.INTERACTIVE_SAFE_CATEGORIES
-
-    @classmethod
-    def skip_reason(cls, option: str) -> str:
-        """Return skip reason or empty string."""
-        return cls.get(option).get("skip_reason", "")  # type: ignore[no-any-return]
-
-    @classmethod
-    def skip_category(cls, option: str) -> str:
-        """Return skip category name."""
-        return cls.get(option)["category"]  # type: ignore[no-any-return]
-
-    @classmethod
-    def safe_options(cls, all_options) -> list:  # type: ignore[no-untyped-def, type-arg]
-        """Return sorted list of options safe for ``--test``."""
-        return sorted(
-            [o for o in all_options if cls.is_safe(o)],
-            key=lambda x: float(x.replace("a", ".1")),
-        )
-
-    @classmethod
-    def interactive_safe_options(cls, all_options) -> list:  # type: ignore[no-untyped-def, type-arg]
-        """Return sorted list of options safe for ``--testinteractive``."""
-        return sorted(
-            [o for o in all_options if cls.is_interactive_safe(o)],
-            key=lambda x: float(x.replace("a", ".1")),
-        )
-
-    @classmethod
-    def unsafe_options(cls, all_options) -> list:  # type: ignore[no-untyped-def, type-arg]
-        """Return sorted list of options NOT safe for ``--test``."""
-        return sorted(
-            [o for o in all_options if not cls.is_safe(o)],
-            key=lambda x: float(x.replace("a", ".1")),
-        )
-
-    @classmethod
-    def wave1_entry_routing_baseline(cls) -> dict[str, str]:
-        """Return a copy of the Wave 1 routing baseline map for deterministic tests."""
-        return dict(cls.WAVE1_ENTRY_ROUTING_BASELINE)
-
-    @classmethod
-    def wave1_safety_classification_baseline(cls) -> dict[str, list[str]]:
-        """Return a copy of the Wave 1 safety classification baseline for deterministic tests."""
-        return {key: list(values) for key, values in cls.WAVE1_SAFETY_CLASSIFICATION_BASELINE.items()}
+# OperationRegistry moved to src/utils/operation_registry.py (1013 SC-001 position 13)
 
 
 def _systematic_test_build_safe_list(
